src/auth/auth.controller.ts
import { Body, Controller, Post } from '@nestjs/common';
import { CreateUserDto, LoginUserDto } from '../users/infra/dto/user.dto';
import { AuthService } from './auth.service';

@Controller('auth')
export class AuthController {
  constructor(private readonly authService: AuthService) {}

  @Post('signin')
  async signIn(@Body() user: LoginUserDto) {
    return await this.authService.signIn(user);
  }

  @Post('signup')
  async signUp(@Body() user: CreateUserDto) {
    return await this.authService.signUp(user);
  }
}
